cleaner way to remove conditionals from modulo (I think)
