+++ devito/ir/iet/scheduler.py
@@ -122,11 +122,10 @@ def iet_make(clusters):
         for j, offs in ispace.sub_iterators.get(k.dim, []):
             modulo = len(offs)
             for n, o in enumerate(filter_ordered(offs)):
-                if j.is_Stepping:
-                    name = "%s%d" % (j.name, n)
-                    vname = Scalar(name=name, dtype=np.int32)
-                    value = (k.dim + o) % modulo
-                    uindices.append(UnboundedIndex(vname, value, value, j, j + o))
+                name = "%s%d" % (j.name, n)
+                vname = Scalar(name=name, dtype=np.int32)
+                value = (k.dim + o) % modulo
+                uindices.append(UnboundedIndex(vname, value, value, j, j + o))
         mapper[k] = k._rebuild(uindices=uindices)
     iet = NestedTransformer(mapper).visit(iet)
