+++ tests/test_dimension.py
@@ -208,13 +208,13 @@ class TestConditionalDimension(object):
         u = TimeFunction(name='u', grid=grid, save=nt)
         assert(grid.time_dim in u.indices)
 
-        # Creates subsampled spatial dimensions and accordine grid
+        # Creates subsampled spatial dimensions and according grid
         dims = tuple([ConditionalDimension(d.name+'sub', parent=d, factor=2)
                       for d in u.grid.dimensions])
         grid2 = Grid((6, 6), dimensions=dims)
         u2 = TimeFunction(name='u2', grid=grid2, save=nt, space_order=2)
         assert(time in u2.indices)
-        print(u2.dx)
+
         eqns = [Eq(u.forward, u + 1.), Eq(u2, u), Eq(u2.forward, u2.dx)]
         op = Operator(eqns, dse="advanced")
         op.apply(time_M=nt-2)
